databae fixing

+++ bd/models.py
@@ -1,5 +1,5 @@
 from sqlalchemy.orm import DeclarativeBase
-from sqlalchemy import  Column, Integer, String, Boolean
+from sqlalchemy import  Column, Integer, String, Boolean, BigInteger
 from bd.connections import pgsql_engine
 
 class Base(DeclarativeBase): pass
@@ -8,9 +8,9 @@ class Message(Base):
     __tablename__ = 'messages'
 
     message_id = Column(Integer, primary_key=True, index=True)
-    chat_id = Column(Integer, primary_key=True)
+    chat_id = Column(BigInteger, primary_key=True)
     from_user = Column(Integer, index=True)
-    from_chat = Column(Integer)
+    from_chat = Column(BigInteger)
     origin_id = Column(Integer)
     type = Column(String)
     answer_id = Column(String)
